Update to API and APP
APP: An issue caused docker to use 100% of one core. We have switched to
meteor to fix this issue.

API: Updated to add the ability to use plugins with go 1.8. However, in
order to use plugins the main app and plugins need to be compiled on the
same host. We now use golang:1.8.1 image and build the api-server and
any plugin in the plugins directory.